# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Сервис купли-продажи
  description: API для сервиса купли-продажи товаров (аналог Avito)
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/admin/users:
    post:
      tags:
        - Администратор
      summary: Создание пользователя администратором
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    get:
      tags:
        - Администратор
      summary: Получение информации о пользователях
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/admin/users/{id}:
    delete:
      tags:
        - Администратор
      summary: Удаление пользователя администратором
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пользователь удален
  /api/admin/users/{id}/role:
    put:
      tags:
        - Администратор
      summary: Изменение роли пользователя администратором
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
      responses:
        '200':
          description: Роль пользователя обновлена
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        password:
          type: string
...
    RoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [USER, ADMIN, WORKER, MANAGER]
...